+++ src/Facades/Slack.php
@@ -13,6 +13,7 @@ use Illuminate\Support\Facades\Facade;
  * @method static getUserProfile(string $user_id)
  * @method static sendMessageUsingWebhook(\Slack\Objects\SlackMessage $message, string $webhookUrl)
  * @method static sendMessage(\Slack\Objects\SlackMessage $message)
+ * @method static deleteMessage(string $channel_id, string $message_ts)
  * @method static openView(\Slack\Objects\SlackView $view, string $trigger_id)
  * @method static updateView(\Slack\Objects\SlackView $view, string $view_id = null, string $external_id = null, string $hash = null)
  * @method static pushView(\Slack\Objects\SlackView $view, string $trigger_id)